requirements.txt
feedparser==6.0.11
requests==2.32.3
deep-translator==1.11.4
slack_sdk==3.27.1
notion-client==2.2.1
python-dotenv==1.0.1
openai==1.35.7 